+++ src/ReadoutEquipment.h
@@ -85,8 +85,11 @@ class ReadoutEquipment {
 
   // Counter values, updated at runtime
   std::vector<CounterStats> equipmentStats;
-   
-  
+  std::vector<CounterValue> equipmentStatsLast;
+     
+  double cfgConsoleStatsUpdateTime=0; // number of seconds between regular printing of statistics on console (if zero, only on stop)
+  AliceO2::Common::Timer consoleStatsTimer; // timer to keep track of elapsed time between console statistics updates
+
   AliceO2::Common::Timer clk;
   AliceO2::Common::Timer clk0;